Move isInFlight view into the focus panel
